requirements.txt
Flask==3.0.3
gunicorn==23.0.0
google-cloud-aiplatform==1.72.0